Refactor the views into category files
