--- nominate/views.py
@@ -1,241 +1,0 @@
-import functools
-from collections.abc import Iterable
-from typing import Any
-
-from django.contrib import messages
-from django.contrib.auth.decorators import login_required
-from django.core.exceptions import PermissionDenied
-from django.http import HttpRequest, HttpResponse
-from django.shortcuts import get_object_or_404, redirect, render
-from django.urls import resolve, reverse
-from django.utils.decorators import method_decorator
-from django.views.generic import DetailView, ListView, RedirectView, TemplateView
-from ipware import get_client_ip
-
-from nominate import models
-from nominate.forms import NominationFormset, RankFormset
-
-
-class ElectionView(ListView):
-    model = models.Election
-
-    def get_queryset(self):
-        query_set: Iterable[models.Election] = super().get_queryset()
-
-        # annotate our elections with some info
-        for election in query_set:
-            election.is_open_for_user = election.is_open_for(self.request.user)
-            election.user_state = election.describe_state(user=self.request.user)
-            election.user_pretty_state = election.pretty_state(user=self.request.user)
-
-        return query_set
-
-
-class ElectionModeView(RedirectView):
-    permanent = False
-    query_string = True
-
-    def get_redirect_url(self, *args: Any, **kwargs: Any) -> str | None:
-        r = resolve(self.request.path)
-        election = get_object_or_404(models.Election, slug=kwargs.get("election_id"))
-        if election.user_can_nominate(self.request.user):
-            return reverse(
-                f"{r.namespace}:nominate", kwargs={"election_id": election.slug}
-            )
-
-        if election.user_can_vote(self.request.user):
-            return reverse(f"{r.namespace}:vote", kwargs={"election_id": election.slug})
-
-        return reverse(f"{r.namespace}:closed", kwargs={"election_id": election.slug})
-
-
-class ClosedElectionView(DetailView):
-    model = models.Election
-    slug_url_kwarg = "election_id"
-    template_name_suffix = "_closed"
-
-
-class NominatorView(TemplateView):
-    @method_decorator(login_required)
-    def dispatch(self, *args, **kwargs):
-        return super().dispatch(*args, **kwargs)
-
-    @functools.lru_cache
-    def election(self):
-        return get_object_or_404(models.Election, slug=self.kwargs.get("election_id"))
-
-    @functools.lru_cache
-    def categories(self):
-        return models.Category.objects.filter(election=self.election())
-
-    def get_context_data(self, **kwargs):
-        ctx = {
-            "election": self.election,
-            "categories": self.categories,
-            "profile": self.profile,
-        }
-        ctx.update(super().get_context_data(**kwargs))
-        return ctx
-
-    @functools.lru_cache
-    def profile(self):
-        try:
-            profile = self.request.user.convention_profile
-        except models.NominatingMemberProfile.DoesNotExist:
-            raise PermissionDenied()
-
-        return profile
-
-
-class NominationView(NominatorView):
-    template_name = "nominate/nominate.html"
-
-    def build_ballot_forms(self, data=None):
-        args = [] if data is None else [data]
-        return {
-            category: NominationFormset(
-                *args,
-                form_kwargs={"category": category},
-                queryset=models.Nomination.objects.filter(
-                    category=category, nominator=self.profile()
-                ),
-                prefix=str(category.id),
-            )
-            for category in self.categories()
-        }
-
-    def get_context_data(self, **kwargs):
-        formsets = kwargs.pop("formsets", None)
-        if formsets is None:
-            formsets = self.build_ballot_forms()
-        ctx = {"formsets": formsets}
-        ctx.update(super().get_context_data(**kwargs))
-        return ctx
-
-    def get(self, request: HttpRequest, *args, **kwargs):
-        if not self.election().user_can_nominate(request.user):
-            self.template_name = "nominate/election_closed.html"
-            self.extra_context = {"object": self.election()}
-
-        return super().get(request, *args, **kwargs)
-
-    def post(self, request: HttpRequest, *args, **kwargs):
-        if not self.election().user_can_nominate(request.user):
-            messages.error(
-                request, f"You do not have nominating rights for {self.election()}"
-            )
-            r = resolve(self.request.path)
-            return redirect(f"{r.namespace}:index")
-
-        profile = self.profile()
-        had_errors = False
-        client_ip_address, _ = get_client_ip(request=request)
-        formsets = {
-            category: NominationFormset(
-                request.POST,
-                request.FILES,
-                form_kwargs={"category": category},
-                prefix=str(category.id),
-            )
-            for category in self.categories()
-        }
-
-        for category, formset in formsets.items():
-            if formset.is_valid():
-                for nomination_record in formset.save(commit=False):
-                    nomination_record.category = category
-                    nomination_record.nominator = profile
-                    nomination_record.nomination_ip_address = client_ip_address
-                    nomination_record.save()
-
-                for deleted in formset.deleted_objects:
-                    deleted.delete()
-            else:
-                had_errors = True
-
-        if not had_errors:
-            messages.success(request, "Your set of nominations was saved")
-            r = resolve(self.request.path)
-            return redirect(
-                f"{r.namespace}:nominate", election_id=self.kwargs.get("election_id")
-            )
-        else:
-            messages.warning(request, "Something wasn't quite right with your ballot")
-            return self.render_to_response(self.get_context_data(formsets=formsets))
-
-
-class VoteView(NominatorView):
-    template_name = "nominate/vote.html"
-
-    def build_ballot_forms(self, data=None):
-        args = [] if data is None else [data]
-        return {
-            category: RankFormset(
-                *args,
-                form_kwargs={"category": category},
-                queryset=models.Rank.objects.filter(
-                    finalist__category=category, membership=self.profile()
-                ),
-                prefix=str(category.id),
-            )
-            for category in self.categories()
-        }
-
-    def get_context_data(self, **kwargs):
-        formsets = kwargs.pop("formsets", None)
-        if formsets is None:
-            formsets = self.build_ballot_forms()
-        ctx = {"formsets": formsets}
-        ctx.update(super().get_context_data(**kwargs))
-        return ctx
-
-    def get(self, request: HttpRequest, *args, **kwargs):
-        if not self.election().user_can_vote(request.user):
-            self.template_name = "nominate/election_closed.html"
-
-        return super().get(request, *args, **kwargs)
-
-    def post(self, request: HttpRequest, *args, **kwargs):
-        if not self.election().user_can_nominate(request.user):
-            messages.error(
-                request, f"You do not have nominating rights for {self.election()}"
-            )
-            r = resolve(self.request.path)
-            return redirect(f"{r.namespace}:index")
-
-        profile = self.profile()
-        had_errors = False
-        client_ip_address, _ = get_client_ip(request=request)
-        formsets = {
-            category: NominationFormset(
-                request.POST,
-                request.FILES,
-                form_kwargs={"category": category},
-                prefix=str(category.id),
-            )
-            for category in self.categories()
-        }
-
-        for category, formset in formsets.items():
-            if formset.is_valid():
-                for nomination_record in formset.save(commit=False):
-                    nomination_record.category = category
-                    nomination_record.nominator = profile
-                    nomination_record.nomination_ip_address = client_ip_address
-                    nomination_record.save()
-            else:
-                had_errors = True
-
-        if not had_errors:
-            messages.success(request, "Your set of nominations was saved")
-            r = resolve(self.request.path)
-            return redirect(
-                f"{r.namespace}:nominate", election_id=self.kwargs.get("election_id")
-            )
-        else:
-            messages.warning(request, "Something wasn't quite right with your ballot")
-            return self.render_to_response(self.get_context_data(formsets=formsets))
-
-
-def access_denied(request: HttpRequest, *args, **kwargs) -> HttpResponse:
-    return render(request, "nominate/forbidden.html")
