+++ nominate/views/__init__.py
@@ -1,0 +1,5 @@
+# ruff: noqa: F401
+from .base import access_denied
+from .election import ClosedElectionView, ElectionModeView, ElectionView
+from .nominate import NominationView
+from .vote import VoteView
